Add a class to use a keystoneauth session too
This commit adds a translation layer for tempest clients to use a
keystoneauth session. It adds a "auth_provider" class that takes in
a Session object from keystoneauth and uses that to behave in a manner
that tempest-lib's rest client is expecting.

Since tempest-lib's auth provider classes weren't really written in a
manner that was expecting things to be used with things that behave
differently and keystone auth is kind of a parallel to tempest-lib's
rest client and auth providers this is sort of hacky. But, I think
this will work.

Disclamer not tested at all, it's just a starting point