Fix v3 auth_url building
We were previously assuming the auth_url provided by os-client-config
was the versioned keystone endpoing. However in some cases, like
with vexxhost, this isn't the case. This commit adds logic to the
v3 side to construct a versioned endpoint from an unversioned one,
which is what the tempest.lib clients expect